Extract shared temp MIDI test fixture helper
